TST: Update test for github and add test for the discovery
