+++ loktar/src/tests/test_scm_github.py
@@ -1,114 +1,10 @@
-from github import GithubException
-from mock import MagicMock
 import pytest
+
+from conftest import FakeGithubRepo
 
 from loktar.exceptions import SCMError
 from loktar.scm import fetch_github_file
 from loktar.scm import Github
-
-
-class File(object):
-
-        def __init__(self, name):
-            self.filename = name
-
-
-class Commit(object):
-        files = [File("toto"), File("toto"), File("lulu")]
-        sha = 'sha'
-
-        def __init__(self, *args, **kwargs):
-
-            class Commit_(object):
-                message = 'message'
-
-            self.commit = Commit_()
-
-        def get_statuses(self):
-            return ['status1', 'status2']
-
-
-class PullRequestPart(object):
-        ref = 'branch'
-        sha = 'commitsha'
-
-
-class IssueComment(object):
-        body = 'existing comment'
-
-
-class PullRequest(object):
-        head = PullRequestPart()
-        number = 565
-
-        def __init__(self, *args, **kwargs):
-            pass
-
-        def get_commits(self):
-            # Mocking a paginated list
-            m = MagicMock(reversed=[Commit(), Commit()])
-
-            def one_element():
-                yield Commit()
-                yield Commit()
-            m.__iter__.side_effect = one_element
-            return m
-
-        def get_issue_comments(self):
-            return [IssueComment()]
-
-        def create_issue_comment(self, comment):
-            return comment
-
-
-class Repository(object):
-        def __init__(self, fail):
-            self.fail = fail
-
-        def get_pulls(self, state=None):
-            return [
-                PullRequest(),
-                PullRequest()
-            ]
-
-        def get_pull(self, state=None):
-            return PullRequest()
-
-        def get_commit(self, sha):
-            return Commit()
-
-        def commit(self, sha):
-            return Commit()
-
-        def create_git_tag_and_release(self, tag_name, tag_message, release_name, patch_note, commit_id, type_object="commit"):
-            if self.fail["status"] == 201 and self.fail["exc"] is False:
-                return {
-                    "status": "201 Created"
-                }
-            else:
-                if self.fail["status"] == 400:
-                    return {
-                        "status": "400 blabla"
-                    }
-                else:
-                    raise GithubException("", "")
-
-
-class fakePR(object):
-        def __init__(self, fail):
-            self.fail = fail
-
-        def get_organization(self, *args, **kwargs):
-            mock_get_repo = MagicMock()
-            mock_get_repo.get_repo.return_value = Repository(self.fail)
-            return mock_get_repo
-
-
-@pytest.fixture
-def mock_github_obj(mocker):
-    github_mock = mocker.patch('loktar.scm.GitHub')
-
-    return github_mock
 
 
 @pytest.mark.parametrize("fail", [
@@ -119,29 +15,41 @@ def mock_github_obj(mocker):
 def test_github(mock_github_obj, fail):
 
     def fake(fail, *args, **kwargs):
-        return fakePR(fail)
+        return FakeGithubRepo(fail)
 
     mock_github = mock_github_obj
     mock_github.return_value = fake(fail)
+    mock_github.self.pull_requests_cache.return_value = {42}
 
     scm = Github("toto", "titi")
 
     scm.search_pull_request_id("branch")
-    pr_info = scm.get_pull_request(42)
-    assert pr_info.head.ref == "branch"
-    assert pr_info.head.sha == "commitsha"
-    assert pr_info.number == 565
-    assert pr_info.get_issue_comments()[0].body == "existing comment"
+    if fail["exc"] is False:
+        pr_info = scm.get_pull_request(42)
+        assert pr_info.head.ref == "branch"
+        assert pr_info.head.sha == "commitsha"
+        assert pr_info.number == 565
+        assert pr_info.get_issue_comments()[0].body == "existing comment"
 
-    commit_messages = scm.get_commit_message_modified_files_on_pull_request(42)
-    scm.get_pull_requests()
-    assert commit_messages == {'message': ['toto', 'toto', 'lulu', 'toto', 'toto', 'lulu']}
-    assert scm.create_pull_request_comment(42, 'comment') == 'comment'
-    assert scm.create_pull_request_comment(42, 'existing comment', check_unique=True) is None
-    assert scm.get_git_branch_from_pull_request(42) == 'branch'
-    last_commit, statuses = scm.get_last_statuses_from_pull_request(42)
-    assert statuses == ['status1', 'status2']
-    scm.get_last_commits_from_pull_request(42)
+        scm.get_pull_request(42, use_cache=False)
+
+        commit_messages = scm.get_commit_message_modified_files_on_pull_request(42)
+        scm.get_pull_requests()
+        scm.get_pull_requests(use_cache=True)
+        assert commit_messages == {'message': ['toto', 'toto', 'lulu', 'toto', 'toto', 'lulu']}
+        assert scm.create_pull_request_comment(42, 'comment') == 'comment'
+        assert scm.create_pull_request_comment(42, 'existing comment', check_unique=True) is None
+        assert scm.get_git_branch_from_pull_request(42) == 'branch'
+        last_commit, statuses = scm.get_last_statuses_from_pull_request(42)
+        assert statuses == ['status1', 'status2']
+        scm.get_last_commits_from_pull_request(42)
+    else:
+        with pytest.raises(SCMError):
+            scm.get_pull_request(42)
+
+        with pytest.raises(SCMError):
+            scm.get_commit_message_modified_files_on_pull_request(42)
+
     if fail["status"] != 201 or fail["exc"] is True:
         with pytest.raises(SCMError) as excinfo:
             scm.set_tag_and_release("tag_name", "tag_message", "release_name", "patch_note", "commit_id")
